harden(elf): size the argv/envp arrays in u64
Audit of unchecked size arithmetic on attacker/device/disk operands: the
high-risk allocation sites are already bounded or checked (ext2, ahci,
mmap, udp, io_uring have guards + selftests), and the count*sizeof alloc
sizes compute in size_t so they cannot wrap. The one u32-width multiply
on an attacker-supplied count was elf_setup_stack (argc+1)*8; it is not
reachable (the argv array alloc fails long before argc hits 2^29) but is
a latent footgun, so compute the size in u64.

Built clean; 89 self-tests PASSED, 0 faults; elf_test passes and boot
execs userspace with argv.

+++ kernel/proc/elf.c
@@ -444,9 +444,12 @@ uint64_t elf_setup_stack(phys_addr_t pml4,
         }
     }
 
-    // Allocate user-VA pointer arrays in kernel heap.
-    uint64_t* argv_va = (uint64_t*)kmalloc((argc + 1) * 8);
-    uint64_t* envp_va = (uint64_t*)kmalloc((envc + 1) * 8);
+    // Allocate user-VA pointer arrays in kernel heap.  Size in u64 so the
+    // (count+1)*8 can never wrap the u32 width (argc/envc are attacker-supplied
+    // via execve; the count can't actually reach 2^29 -- the argv array alloc
+    // fails first -- but computing the size in u64 removes the latent footgun).
+    uint64_t* argv_va = (uint64_t*)kmalloc(((uint64_t)argc + 1) * 8);
+    uint64_t* envp_va = (uint64_t*)kmalloc(((uint64_t)envc + 1) * 8);
     if (!argv_va || !envp_va) {
         if (argv_va) kfree(argv_va);
         if (envp_va) kfree(envp_va);
